+++ src/modules/nft/entrypoints/dto.ts
@@ -251,7 +251,7 @@ export class GetNftTokenUriBody {
     required: false,
   })
   @Transform(({ value }) => value && JSON.parse(value))
-  properties?: any;
+  properties?: [{ [key: string]: string }];
 
   @IsArray()
   @IsOptional()
@@ -319,6 +319,55 @@ export class GetMyNftsResponse {
   updatedAt: Date;
 }
 
+export class GetUserNftsResponse {
+  @ApiProperty()
+  id: number;
+
+  @ApiProperty({ enum: NftSource, examples: [NftSource.UNIVERSE, NftSource.SCRAPER] })
+  source: string;
+
+  @ApiProperty({ example: '0x0000000000000000000000000000000000000000000000000000000000000000' })
+  txHash: string;
+
+  @ApiProperty()
+  name: string;
+
+  @ApiProperty()
+  description: string;
+
+  @ApiProperty()
+  tokenId: string;
+
+  @ApiProperty({ examples: ['jpeg', 'png', 'gif', 'webp', 'mp4'], example: 'jpeg' })
+  artworkType: string;
+
+  @ApiProperty()
+  url: string;
+
+  @ApiProperty()
+  optimized_url: string;
+
+  @ApiProperty()
+  thumbnail_url: string;
+
+  @ApiProperty()
+  original_url: string;
+
+  @ApiProperty()
+  tokenUri: string;
+
+  @ApiProperty({ type: JSON })
+  properties: any;
+
+  @ApiProperty({ example: 10 })
+  royalties: number;
+
+  @ApiProperty()
+  createdAt: Date;
+
+  @ApiProperty({ type: Date })
+  updatedAt: Date;
+}
 export class CreateCollectionBody {
   @IsString()
   @Length(1, 32)
@@ -369,6 +418,16 @@ export class EditCollectionParams {
   id: number;
 }
 
+export class GetUserNftsParams {
+  @IsString()
+  @ApiProperty({
+    example: 'username1',
+    description: 'The username of the user',
+    required: true,
+  })
+  username: string;
+}
+
 export class EditMintingCollectionBody {
   @IsString()
   @IsOptional()
@@ -390,11 +449,11 @@ export class DeleteSavedNftParams {
   id: number;
 }
 
-export class GetMyCollectionParams {
-  @IsNumberString()
+export class GetCollectionParams {
+  @IsString()
   @ApiProperty({
-    description: 'The id of the Collection',
-    example: 1,
+    description: 'The address of the Collection',
+    example: '0x0000000000000000000000000',
   })
-  id: number;
+  address: string;
 }