+++ src/modules/users/users.service.ts
@@ -1,12 +1,13 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { config } from 'node:process';
-import { Repository } from 'typeorm';
+import { Not, Repository } from 'typeorm';
 import { AppConfig } from '../configuration/configuration.service';
 import { S3Service } from '../file-storage/s3.service';
 import { UserInfoDto } from './user.dto';
 import { User } from './user.entity';
 import { UserNotFoundException } from './service-layer/exceptions/UserNotFoundException';
+import { DuplicateUniversePageUrlException } from './service-layer/exceptions/DuplicateUniversePageUrlException';
 
 @Injectable()
 export class UsersService {
@@ -31,7 +32,13 @@ export class UsersService {
   }
 
   async saveProfileInfo(userInfoDto: UserInfoDto, user: any) {
-    const userDb = await this.usersRepository.findOne({ where: { address: user.address } });
+    const [userDb, duplicateUrlUser] = await Promise.all([
+      this.usersRepository.findOne({ where: { address: user.address } }),
+      this.usersRepository.findOne({
+        where: { universePageUrl: userInfoDto.universePageUrl, address: Not(user.address) },
+      }),
+    ]);
+
     if (!userDb) {
       throw new HttpException(
         {
@@ -42,6 +49,10 @@ export class UsersService {
       );
     }
 
+    if (duplicateUrlUser) {
+      throw new DuplicateUniversePageUrlException();
+    }
+
     userDb.displayName = userInfoDto.displayName;
     userDb.universePageUrl = userInfoDto.universePageUrl;
     userDb.about = userInfoDto.about;
@@ -50,7 +61,6 @@ export class UsersService {
 
     await this.usersRepository.save(userDb);
   }
-
   async uploadProfileImage(file: Express.Multer.File, user: any) {
     try {
       const userDb = await this.usersRepository.findOne({ where: { address: user.address } });
@@ -104,6 +114,16 @@ export class UsersService {
 
   async getProfileInfo(address: string) {
     return await this.usersRepository.findOne({ where: { address: address, isActive: true } });
+  }
+
+  async getByUsername(username: string) {
+    const user = await this.usersRepository.findOne({ where: { universePageUrl: username, isActive: true } });
+
+    if (!user) {
+      throw new UserNotFoundException();
+    }
+
+    return user;
   }
 
   async getById(id: number, validate = true) {